+++ src/Autoload.php
@@ -10,3 +10,4 @@ require_once 'Expectations/Database.php';
 require_once 'Expectations/Models.php';
 require_once 'Expectations/Response.php';
 require_once 'Expectations/Storage.php';
+require_once 'Expectations/Time.php';