+++ src/Expectations/Response.php
@@ -317,6 +317,54 @@ expect()->extend(
 );
 
 expect()->extend(
+    'toHaveJsonStructure',
+    /**
+     * Assert that the response has a given JSON structure.
+     */
+    function (array $structure = null, array $responseData = null): Expectation {
+        /** @var TestResponse $response */
+        $response = $this->value;
+
+        $response->assertJsonStructure($structure, $responseData);
+
+        return $this;
+    }
+);
+
+expect()->extend(
+    'toHaveJsonPath',
+    /**
+     * Assert that the expected value and type exists at the given path in the response.
+     *
+     * @param mixed $expect
+     */
+    function (string $path, $expect): Expectation {
+        /** @var TestResponse $response */
+        $response = $this->value;
+
+        $response->assertJsonPath($path, $expect);
+
+        return $this;
+    }
+);
+
+expect()->extend(
+    'toHaveJsonValidationErrors',
+    /**
+     * Assert that the response has the given JSON validation errors.
+     *
+     * @param string|array $errors
+     */
+    function ($errors = null, string $responseKey = 'errors'): Expectation {
+        /** @var TestResponse $response */
+        $response = $this->value;
+        $response->assertJsonValidationErrors($errors, $responseKey);
+
+        return $this;
+    },
+);
+
+expect()->extend(
     'toHaveValid',
     /**
      * Assert that the response doesn't have the given validation error keys.
@@ -362,4 +410,18 @@ expect()->extend(
 
         return $this;
     }
+);
+
+expect()->extend(
+    'toHaveLocation',
+    /**
+     * Assert that the current location header matches the given URI.
+     */
+    function (string $uri): Expectation {
+        /** @var TestResponse $response */
+        $response = $this->value;
+        $response->assertLocation($uri);
+
+        return $this;
+    },
 );